+++ jwst/tweakreg/tweakreg_step.py
@@ -10,8 +10,8 @@ from astropy.table import Table
 from astropy import units as u
 from astropy.coordinates import SkyCoord
 from tweakwcs.imalign import align_wcs
-from tweakwcs.tpwcs import JWSTgWCS
-from tweakwcs.matchutils import TPMatch
+from tweakwcs.correctors import JWSTWCSCorrector
+from tweakwcs.matchutils import XYXYMatch
 
 # LOCAL
 from ..stpipe import Step
@@ -201,7 +201,7 @@ class TweakRegStep(Step):
             self.log.info('')
 
             # align images:
-            tpmatch = TPMatch(
+            xyxymatch = XYXYMatch(
                 searchrad=self.searchrad,
                 separation=self.separation,
                 use2dhist=self.use2dhist,
@@ -217,7 +217,7 @@ class TweakRegStep(Step):
                     enforce_user_order=self.enforce_user_order,
                     expand_refcat=self.expand_refcat,
                     minobj=self.minobj,
-                    match=tpmatch,
+                    match=xyxymatch,
                     fitgeom=self.fitgeometry,
                     nclip=self.nclip,
                     sigma=(self.sigma, 'rmse')
@@ -325,7 +325,7 @@ class TweakRegStep(Step):
                 # Update to separation needed to prevent confusion of sources
                 # from overlapping images where centering is not consistent or
                 # for the possibility that errors still exist in relative overlap.
-                tpmatch_gaia = TPMatch(
+                xyxymatch_gaia = XYXYMatch(
                     searchrad=self.abs_searchrad,
                     separation=self.abs_separation,
                     use2dhist=self.abs_use2dhist,
@@ -351,7 +351,7 @@ class TweakRegStep(Step):
                     enforce_user_order=True,
                     expand_refcat=False,
                     minobj=self.abs_minobj,
-                    match=tpmatch_gaia,
+                    match=xyxymatch_gaia,
                     fitgeom=self.abs_fitgeometry,
                     nclip=self.abs_nclip,
                     sigma=(self.abs_sigma, 'rmse')
@@ -434,7 +434,7 @@ class TweakRegStep(Step):
 
         # create WCSImageCatalog object:
         refang = image_model.meta.wcsinfo.instance
-        im = JWSTgWCS(
+        im = JWSTWCSCorrector(
             wcs=image_model.meta.wcs,
             wcsinfo={'roll_ref': refang['roll_ref'],
                      'v2_ref': refang['v2_ref'],